requirements.txt
cohere
google-genai
python-dotenv
rich
openai
duckduckgo-search
selenium
webdriver-manager
mtranslate
pygame
edge_tts
AppOpener
pywhatkit
keyboard
PyQt5
requests
beautifulsoup4
Pillow